+++ src/explorer.rs
@@ -1,54 +1,58 @@
 use crate::app::{DirectoryBuffer, ExplorerConfig, ExternalMsg, InternalMsg, MsgIn, Node, Task};
+use anyhow::Result;
 use std::fs;
 use std::path::PathBuf;
 use std::sync::mpsc::Sender;
 use std::thread;
 
-pub fn explore(
+pub fn explore_sync(
+    config: ExplorerConfig,
+    parent: String,
+    focused_path: Option<String>,
+) -> Result<DirectoryBuffer> {
+    let path = PathBuf::from(&parent);
+
+    let dirs = fs::read_dir(&path)?;
+    let mut nodes = dirs
+        .filter_map(|d| {
+            d.ok().map(|e| {
+                e.path()
+                    .file_name()
+                    .map(|n| n.to_string_lossy().to_string())
+                    .unwrap_or_default()
+            })
+        })
+        .map(|name| Node::new(parent.clone(), name))
+        .filter(|n| config.filter(n))
+        .collect::<Vec<Node>>();
+
+    nodes.sort_by(|a, b| config.sort(a, b));
+
+    let focus_index = if let Some(focus) = focused_path {
+        nodes
+            .iter()
+            .enumerate()
+            .find(|(_, n)| n.relative_path() == &focus)
+            .map(|(i, _)| i)
+            .unwrap_or(0)
+    } else {
+        0
+    };
+
+    Ok(DirectoryBuffer::new(parent, nodes, focus_index))
+}
+
+pub fn explore_async(
     config: ExplorerConfig,
     parent: String,
     focused_path: Option<String>,
     tx: Sender<Task>,
 ) {
-    let path = PathBuf::from(&parent);
-    let path_cloned = path.clone();
-    let tx_cloned = tx.clone();
-    let config_cloned = config.clone();
-
     thread::spawn(move || {
-        fs::read_dir(&path)
-            .map(|dirs| {
-                let mut nodes = dirs
-                    .filter_map(|d| {
-                        d.ok().map(|e| {
-                            e.path()
-                                .file_name()
-                                .map(|n| n.to_string_lossy().to_string())
-                                .unwrap_or_default()
-                        })
-                    })
-                    .map(|name| Node::new(parent.clone(), name))
-                    .filter(|n| config.filter(n))
-                    .collect::<Vec<Node>>();
-                nodes.sort_by(|a, b| config.sort(a, b));
-                nodes
-            })
-            .map(|nodes| {
-                let focus_index = if let Some(focus) = focused_path {
-                    nodes
-                        .iter()
-                        .enumerate()
-                        .find(|(_, n)| n.relative_path() == &focus)
-                        .map(|(i, _)| i)
-                        .unwrap_or(0)
-                } else {
-                    0
-                };
-
-                let dir = DirectoryBuffer::new(parent.clone(), nodes, focus_index);
-
+        explore_sync(config, parent.clone(), focused_path)
+            .map(|dirbuf| {
                 tx.send(Task::new(
-                    MsgIn::Internal(InternalMsg::AddDirectory(parent, dir)),
+                    MsgIn::Internal(InternalMsg::AddDirectory(parent, dirbuf)),
                     None,
                 ))
                 .unwrap_or_default();
@@ -61,15 +65,22 @@ pub fn explore(
                 .unwrap_or_default();
             })
     });
+}
 
-    if let Some(grand_parent) = path_cloned.parent() {
-        explore(
-            config_cloned,
+pub fn explore_recursive_async(
+    config: ExplorerConfig,
+    parent: String,
+    focused_path: Option<String>,
+    tx: Sender<Task>,
+) {
+    let path = PathBuf::from(&parent);
+    explore_async(config.clone(), parent, focused_path, tx.clone());
+    if let Some(grand_parent) = path.parent() {
+        explore_recursive_async(
+            config,
             grand_parent.to_string_lossy().to_string(),
-            path_cloned
-                .file_name()
-                .map(|f| f.to_string_lossy().to_string()),
-            tx_cloned,
+            path.file_name().map(|f| f.to_string_lossy().to_string()),
+            tx,
         );
     }
 }
