+++ src/ui.rs
@@ -1,7 +1,6 @@
 use crate::app;
 use crate::app::HelpMenuLine;
 use crate::app::{Node, ResolvedNode};
-use crate::config::Layout;
 use crate::config::PanelUiConfig;
 use handlebars::Handlebars;
 use indexmap::IndexSet;
@@ -21,6 +20,115 @@ use tui::Frame;
 lazy_static! {
     pub static ref NO_COLOR: bool = env::var("NO_COLOR").ok().map(|_| true).unwrap_or(false);
     pub static ref DEFAULT_STYLE: TuiStyle = TuiStyle::default();
+}
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct LayoutOptions {
+    #[serde(default)]
+    margin: Option<u16>,
+    #[serde(default)]
+    horizontal_margin: Option<u16>,
+    #[serde(default)]
+    vertical_margin: Option<u16>,
+    #[serde(default)]
+    constraints: Option<Vec<Constraint>>,
+}
+
+impl LayoutOptions {
+    pub fn extend(mut self, other: Self) -> Self {
+        self.margin = other.margin.or(self.margin);
+        self.horizontal_margin = other.horizontal_margin.or(self.horizontal_margin);
+        self.vertical_margin = other.vertical_margin.or(self.vertical_margin);
+        self.constraints = other.constraints.or(self.constraints);
+        self
+    }
+
+    /// Get a reference to the layout options's constraints.
+    pub fn constraints(&self) -> &Option<Vec<Constraint>> {
+        &self.constraints
+    }
+
+    /// Get a reference to the layout options's margin.
+    pub fn margin(&self) -> Option<u16> {
+        self.margin
+    }
+
+    /// Get a reference to the layout options's horizontal margin.
+    pub fn horizontal_margin(&self) -> Option<u16> {
+        self.horizontal_margin
+    }
+
+    /// Get a reference to the layout options's vertical margin.
+    pub fn vertical_margin(&self) -> Option<u16> {
+        self.vertical_margin
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub enum Layout {
+    Nothing(Option<PanelUiConfig>),
+    Table(Option<PanelUiConfig>),
+    InputAndLogs(Option<PanelUiConfig>),
+    Selection(Option<PanelUiConfig>),
+    HelpMenu(Option<PanelUiConfig>),
+    SortAndFilter(Option<PanelUiConfig>),
+    Horizontal {
+        config: LayoutOptions,
+        splits: Vec<Layout>,
+    },
+    Vertical {
+        config: LayoutOptions,
+        splits: Vec<Layout>,
+    },
+}
+
+impl Default for Layout {
+    fn default() -> Self {
+        Self::Nothing(Default::default())
+    }
+}
+
+impl Layout {
+    pub fn extend(self, other: Self) -> Self {
+        match (self, other) {
+            (s, Self::Nothing(_)) => s,
+            (Self::Table(s), Self::Table(o)) => Self::Table(o.or(s)),
+            (Self::InputAndLogs(s), Self::InputAndLogs(o)) => Self::InputAndLogs(o.or(s)),
+            (Self::Selection(s), Self::Selection(o)) => Self::Selection(o.or(s)),
+            (Self::HelpMenu(s), Self::HelpMenu(o)) => Self::HelpMenu(o.or(s)),
+            (Self::SortAndFilter(s), Self::SortAndFilter(o)) => Self::SortAndFilter(o.or(s)),
+            (
+                Self::Horizontal {
+                    config: sconfig,
+                    splits: _,
+                },
+                Self::Horizontal {
+                    config: oconfig,
+                    splits: osplits,
+                },
+            ) => Self::Horizontal {
+                config: sconfig.extend(oconfig),
+                splits: osplits,
+            },
+
+            (
+                Self::Vertical {
+                    config: sconfig,
+                    splits: _,
+                },
+                Self::Vertical {
+                    config: oconfig,
+                    splits: osplits,
+                },
+            ) => Self::Vertical {
+                config: sconfig.extend(oconfig),
+                splits: osplits,
+            },
+            (_, other) => other,
+        }
+    }
 }
 
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Serialize, Deserialize)]
@@ -114,8 +222,7 @@ impl Into<TuiStyle> for Style {
                         .unwrap_or_default()
                         .into_iter()
                         .map(|m| m.bits())
-                        .reduce(|a, b| (a ^ b))
-                        .unwrap_or_else(|| TuiModifier::empty().bits()),
+                        .fold(0, |a, b| (a ^ b)),
                 ),
 
                 sub_modifier: TuiModifier::from_bits_truncate(
@@ -123,10 +230,63 @@ impl Into<TuiStyle> for Style {
                         .unwrap_or_default()
                         .into_iter()
                         .map(|m| m.bits())
-                        .reduce(|a, b| (a ^ b))
-                        .unwrap_or_else(|| TuiModifier::empty().bits()),
+                        .fold(0, |a, b| (a ^ b)),
                 ),
             }
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub enum Constraint {
+    Percentage(u16),
+    Ratio(u32, u32),
+    Length(u16),
+    LengthLessThanScreenHeight(u16),
+    LengthLessThanScreenWidth(u16),
+    LengthLessThanLayoutHeight(u16),
+    LengthLessThanLayoutWidth(u16),
+    Max(u16),
+    MaxLessThanScreenHeight(u16),
+    MaxLessThanScreenWidth(u16),
+    MaxLessThanLayoutHeight(u16),
+    MaxthLessThanLayoutWidth(u16),
+    Min(u16),
+    MinLessThanScreenHeight(u16),
+    MinLessThanScreenWidth(u16),
+    MinLessThanLayoutHeight(u16),
+    MinLessThanLayoutWidth(u16),
+}
+
+impl Constraint {
+    pub fn to_tui(self, screen_size: Rect, layout_size: Rect) -> TuiConstraint {
+        match self {
+            Self::Percentage(n) => TuiConstraint::Percentage(n),
+            Self::Ratio(x, y) => TuiConstraint::Ratio(x, y),
+            Self::Length(n) => TuiConstraint::Length(n),
+            Self::LengthLessThanScreenHeight(n) => {
+                TuiConstraint::Length(screen_size.height.max(n) - n)
+            }
+            Self::LengthLessThanScreenWidth(n) => {
+                TuiConstraint::Length(screen_size.width.max(n) - n)
+            }
+            Self::LengthLessThanLayoutHeight(n) => {
+                TuiConstraint::Length(layout_size.height.max(n) - n)
+            }
+            Self::LengthLessThanLayoutWidth(n) => {
+                TuiConstraint::Length(layout_size.width.max(n) - n)
+            }
+            Self::Max(n) => TuiConstraint::Max(n),
+            Self::MaxLessThanScreenHeight(n) => TuiConstraint::Max(screen_size.height.max(n) - n),
+            Self::MaxLessThanScreenWidth(n) => TuiConstraint::Max(screen_size.width.max(n) - n),
+            Self::MaxLessThanLayoutHeight(n) => TuiConstraint::Max(layout_size.height.max(n) - n),
+            Self::MaxthLessThanLayoutWidth(n) => TuiConstraint::Max(layout_size.width.max(n) - n),
+            Self::Min(n) => TuiConstraint::Min(n),
+            Self::MinLessThanScreenHeight(n) => TuiConstraint::Min(screen_size.height.max(n) - n),
+            Self::MinLessThanScreenWidth(n) => TuiConstraint::Min(screen_size.width.max(n) - n),
+            Self::MinLessThanLayoutHeight(n) => TuiConstraint::Min(layout_size.height.max(n) - n),
+            Self::MinLessThanLayoutWidth(n) => TuiConstraint::Min(layout_size.width.max(n) - n),
         }
     }
 }
@@ -242,8 +402,7 @@ fn block<'a>(config: PanelUiConfig, default_title: String) -> Block<'a> {
                 .unwrap_or_default()
                 .iter()
                 .map(|b| b.bits())
-                .reduce(|a, b| (a ^ b))
-                .unwrap_or_else(|| TuiBorders::NONE.bits()),
+                .fold(0, |a, b| (a ^ b)),
         ))
         .title(Span::styled(
             config.title().format().clone().unwrap_or(default_title),
